Create mdextract.0.0.72.py
Just tidied it up so it outputs a functioning csv file that the front-end guys can start using for their testing (albeit with very dodgily named variables).
HOWEVER! It currently only works for image files and pdf files AFAIK (and only if they have their normal extensions at that).
.py and .txt files will return a 0, all other files will return an error AFAIK.
We need to do a lot more work on cleaning this program up and making it more functional.